+++ liquidfun/cppbuild.sh
@@ -35,16 +35,16 @@ case $PLATFORM in
         ;;
     windows-x86)
         cd liquidfun/Box2D
-        CXXFLAGS="/Wv:17" $CMAKE -G "Visual Studio 14 2015" -DBOX2D_INSTALL=ON -DCMAKE_INSTALL_PREFIX=$INSTALL_PATH -DBOX2D_BUILD_SHARED=ON -DBOX2D_BUILD_EXAMPLES=OFF -DBOX2D_BUILD_UNITTESTS=OFF .
-        PLATFORM= MSBuild.exe Box2D.sln //p:Configuration=Release //maxcpucount:$MAKEJ
+        CXXFLAGS="/Wv:17" $CMAKE -G "Visual Studio 15 2017" -DBOX2D_INSTALL=ON -DCMAKE_INSTALL_PREFIX=$INSTALL_PATH -DBOX2D_BUILD_SHARED=ON -DBOX2D_BUILD_EXAMPLES=OFF -DBOX2D_BUILD_UNITTESTS=OFF .
+        MSBuild.exe Box2D.sln //p:Configuration=Release //p:Platform=Win32 //p:CL_MPCount=$MAKEJ
         cp -r Box2D $INSTALL_PATH/include
         cp Box2D/Release/*.lib $INSTALL_PATH/lib
         cp Box2D/Release/*.dll $INSTALL_PATH/bin
         ;;
     windows-x86_64)
         cd liquidfun/Box2D
-        CXXFLAGS="/Wv:17" $CMAKE -G "Visual Studio 14 2015 Win64" -DBOX2D_INSTALL=ON -DCMAKE_INSTALL_PREFIX=$INSTALL_PATH -DBOX2D_BUILD_SHARED=ON -DBOX2D_BUILD_EXAMPLES=OFF -DBOX2D_BUILD_UNITTESTS=OFF .
-        PLATFORM= MSBuild.exe Box2D.sln //p:Configuration=Release //maxcpucount:$MAKEJ
+        CXXFLAGS="/Wv:17" $CMAKE -G "Visual Studio 15 2017 Win64" -DBOX2D_INSTALL=ON -DCMAKE_INSTALL_PREFIX=$INSTALL_PATH -DBOX2D_BUILD_SHARED=ON -DBOX2D_BUILD_EXAMPLES=OFF -DBOX2D_BUILD_UNITTESTS=OFF .
+        MSBuild.exe Box2D.sln //p:Configuration=Release //p:Platform=x64 //p:CL_MPCount=$MAKEJ
         cp -r Box2D $INSTALL_PATH/include
         cp Box2D/Release/*.lib $INSTALL_PATH/lib
         cp Box2D/Release/*.dll $INSTALL_PATH/bin